pyproject.toml
[tool.poetry]
name = "llm-agent-server"
version = "0.1.0"
description = "A FastAPI server for orchestrating LLM agents"
authors = ["André Lobo Teixeira <andreloboteixeira@gmail.com>"]
readme = "README.md"
packages = [
    { include = "fast_api_server", from = "src" },
    { include = "llm_agents", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = {extras = ["standard"], version = "^0.115.7"}
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.1"
pydantic = "^2.10.6"
prometheus-client = "^0.17.1"
openai = "^1.12.0"
sentence-transformers = "^3.4.1"
numpy = "^2.2.2"
faiss-cpu = "^1.10.0"
python-dotenv = "^1.0.0"
bs4 = "^0.0.2"
requests = "^2.31.0"
google-api-python-client = "^2.108.0"
google-search-results = "^2.4.2"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
ruff = "^0.3.0"  # Fast Python linter, replaces flake8, black, and isort

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto" 